CLaxIntegerをC++17で再定義し、基本型を指定できるようにする
